Backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
flask-jwt-extended
langchain-openai
langchain-core
youtube-transcript-api
langchain-community
langchain-pinecone
nltk 
scikit-learn 
numpy
